+++ sarmfsw.h
@@ -1,8 +1,7 @@
 /*!\file sarmfsw.h
 ** \author SMFSW
-** \date 2017
-** \copyright MIT (c) 2017, SMFSW
-** \brief ARM common headers for projects
+** \copyright MIT (c) 2017-2018, SMFSW
+** \brief sarmfsw ARM common headers
 */
 /****************************************************************/
 #ifndef __SARMFSW_H
@@ -13,28 +12,28 @@
 #endif
 /****************************************************************/
 
-#include "arm_attributes.h"		// Common attributes depending compiler
+#include "arm_attributes.h"		// Common attributes (following compiler)
 #include "arm_typedefs.h"		// Common typedefs
 
 #include "arm_errors.h"			// Common errors
-
 #include "arm_macros.h"			// Common macros
 #include "arm_stdclib.h"		// Common standard c library wrapper macros
 
-#include "arm_cmsis.h"			// HAL & Drivers depending platform
-#include "arm_hal_peripheral.h"	// HAL peripherals includes
+#include "arm_cmsis.h"			// HAL & Drivers (following defined platform)
+#include "arm_hal_peripheral.h"	// HAL peripherals includes (following defined platform)
 
-#include "arm_inlines.h"		// Common inlines (at the end as some needs to access HAL)
+#include "arm_inlines.h"		// Common inlines (after all includes, some inlines needs to access HAL code)
 
 /*!\enum FW_target
 ** \brief Firmware target types
+** \note A variable has to be defined in project using FW_target enum to be meaningful
 **/
 typedef enum PACK__ FW_target {
-	DefSpecialTarget = 0,		//!< Special FW target (same as debug, yet)
-	DefDebugTarget,				//!< Debug FW target (default)
-	DefReleaseTarget,			//!< Release FW target (No debug information)
-	DefFUBARTarget,				//!< FUBAR FW target (shall be used only for stress/testing purposes)
-	DefUnknownTarget = 0xFF		//!< Unknown FW target
+	target_Special = 0,		//!< Special FW target
+	target_Debug,			//!< Debug FW target
+	target_Release,			//!< Release FW target (No debug information)
+	target_FUBAR,			//!< FUBAR FW target (shall be used only for stress test / optimization testing purposes)
+	target_Unknown = 0xFF	//!< Unknown FW target
 } FW_target;